requirements.txt
pywin32
pycryptodome